ml8511 done

+++ components/ml8511/include/ml8511.h
@@ -2,8 +2,8 @@
 #ifndef __ML8511_H__
 #define __ML8511_H__
 
-#include "hal/adc_types.h"
+#include "esp_adc/adc_oneshot.h"
 
 void ml8511_init (adc_unit_t unit, adc_channel_t channel);
-void get_uv_intensity(adc_channel_t channel, int *data);
+void get_uv_intensity(adc_channel_t channel, float *data);
 #endif
